back_end/ssup_backend/requirements.txt
django>=4.2.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
mysqlclient>=2.2.0
